vf - End game button seems to be working correctly.
The game will display who the winner is, based on who has the
most amount of cash, we may want to change this later so that we
take into account their properties. A good way to do this, as suggested by Huijun is to force all the players to sell their properties when the
game ends.